+++ test/playlistPermutations.js
@@ -1,4 +1,4 @@
-import { configSyncDefaults } from "../src/config.js";
+import { configSyncDefaults } from "./testConfig.js";
 import { deepCopy } from "../src/utils.js";
 
 // ---------- Constants used by the permutations ----------
@@ -20,12 +20,6 @@ export const times = {
 }
 
 // ---------- Config ----------
-// Any overrides to the default that needs to be different for the tests
-const overrides = {
-	shuffleOpenAsPlaylistOption: true
-};
-const configSyncTestDefaults = { ...configSyncDefaults, ...overrides };
-
 // The configSync should always be the default, with these settings being changed
 // This allows us to not have to define something for every setting
 const configSyncModifiers = [
@@ -124,7 +118,7 @@ for (const useCustomApiKeyOption of configSyncModifiers[0]) {
 			// Exclude invalid combinations
 			if (!useCustomApiKeyOption && (customYoutubeApiKey !== null || databaseSharingEnabledOption)) continue;
 
-			let modifiedConfigSync = deepCopy(configSyncTestDefaults);
+			let modifiedConfigSync = deepCopy(configSyncDefaults);
 			modifiedConfigSync.useCustomApiKeyOption = useCustomApiKeyOption;
 			modifiedConfigSync.customYoutubeApiKey = customYoutubeApiKey;
 			modifiedConfigSync.databaseSharingEnabledOption = databaseSharingEnabledOption;
@@ -143,7 +137,7 @@ for (const shuffleOpenInNewTabOption of configSyncModifiers[3]) {
 			if (!shuffleOpenInNewTabOption && shuffleReUseNewTabOption) continue;
 			if (!shuffleReUseNewTabOption && shuffleTabId !== null) continue;
 
-			let modifiedConfigSync = deepCopy(configSyncTestDefaults);
+			let modifiedConfigSync = deepCopy(configSyncDefaults);
 			modifiedConfigSync.shuffleOpenInNewTabOption = shuffleOpenInNewTabOption;
 			modifiedConfigSync.shuffleReUseNewTabOption = shuffleReUseNewTabOption;
 			modifiedConfigSync.shuffleTabId = shuffleTabId;
@@ -157,7 +151,7 @@ configSyncPermutations.openInNewTabPermutations = openInNewTabPermutations;
 // Ignore shorts
 const ignoreShortsPermutations = [];
 for (const shuffleIgnoreShortsOption of configSyncModifiers[6]) {
-	let modifiedConfigSync = deepCopy(configSyncTestDefaults);
+	let modifiedConfigSync = deepCopy(configSyncDefaults);
 	modifiedConfigSync.shuffleIgnoreShortsOption = shuffleIgnoreShortsOption;
 
 	ignoreShortsPermutations.push(modifiedConfigSync);
@@ -168,7 +162,7 @@ configSyncPermutations.ignoreShortsPermutations = ignoreShortsPermutations;
 const openAsPlaylistPermutations = [];
 for (const shuffleOpenAsPlaylistOption of configSyncModifiers[7]) {
 	for (const shuffleNumVideosInPlaylist of configSyncModifiers[8]) {
-		let modifiedConfigSync = deepCopy(configSyncTestDefaults);
+		let modifiedConfigSync = deepCopy(configSyncDefaults);
 		// Always open in a new tab, so we can check the stub
 		modifiedConfigSync.shuffleOpenInNewTabOption = true;
 
@@ -190,7 +184,7 @@ for (const activeOption of configSyncModifiers[9]) {
 		// The allVideosOption and percentageOption do not error out if no value is set
 		if (["allVideosOption", "percentageOption"].includes(activeOption) && channelSettingsPermutation.type === "empty") continue;
 
-		let modifiedConfigSync = deepCopy(configSyncTestDefaults);
+		let modifiedConfigSync = deepCopy(configSyncDefaults);
 		let usedChannelSettingsPermutation = deepCopy(channelSettingsPermutation);
 
 		usedChannelSettingsPermutation.template.activeOption = activeOption;
@@ -574,7 +568,7 @@ export function needsDBInteraction(permutation) {
 }
 
 // Determine whether or not a permutation needs to interact with the YouTube API
-export function needsYTAPIInteraction(permutation, configSync = configSyncTestDefaults) {
+export function needsYTAPIInteraction(permutation, configSync = configSyncDefaults) {
 	const databaseSharing = configSync.databaseSharingEnabledOption;
 	if (databaseSharing) {
 		return (needsDBInteraction(permutation) &&